% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_isoformic.R
\name{as_isoformic}
\alias{as_isoformic}
\title{Convert a SummarizedExperiment to an IsoformicExperiment Object}
\usage{
as_isoformic(se, annot_path, annot_type = c("gff", "annot_db"))
}
\arguments{
\item{se}{A \code{SummarizedExperiment} object to be converted.}

\item{annot_path}{Path to the annotation file. This can be a GFF file or the
path pre-built annotation database created with \verb{[prepare_isoformic_annotation()]}.}
}
\description{
This function converts a \code{SummarizedExperiment} object to an \code{IsoformicExperiment} object.
It extracts the assays, row data, column data, and metadata from the input object
and uses them to create a new \code{IsoformicExperiment} object.
}
