% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_mae_isoformic.R
\name{check_mae_isoformic_is_list}
\alias{check_mae_isoformic_is_list}
\title{Check if Isoformic \code{mae} have a list \code{isoformic} slot}
\usage{
check_mae_isoformic_is_list(mae)
}
\description{
Check if Isoformic \code{mae} have a list \code{isoformic} slot
}
