% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_mae_isoformic.R
\name{check_mae_isoformic_tx_to_gene}
\alias{check_mae_isoformic_tx_to_gene}
\title{Check if Isoformic \code{mae} have a \code{tx_to_gene}}
\usage{
check_mae_isoformic_tx_to_gene(mae)
}
\description{
Check if Isoformic \code{mae} have a \code{tx_to_gene}
}
