% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_utils.R
\name{check_se_metadata_level_type}
\alias{check_se_metadata_level_type}
\title{Check if content of metadata level slot is a string.}
\usage{
check_se_metadata_level_type(se)
}
\description{
Check if content of metadata level slot is a string.
}
