% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContextData-class.R
\name{ContextData}
\alias{ContextData}
\title{ContextData Class}
\usage{
ContextData(
  gene_name = character(0),
  gff_file = character(0),
  txdb = NULL,
  annotation_table = (function (.data = list(), row.names = NULL) 
 {
     if
    (is.null(row.names)) {
         list2DF(.data)
     }
     else {
         out <-
    list2DF(.data, length(row.names))
         attr(out, "row.names") <- row.names
      
      out
     }
 })(),
  annotation_name = character(0),
  assembly_name = character(0),
  ideogram_assembly = character(0),
  organism = character(0),
  orgdb_package = character(0),
  bsgenome_package = character(0),
  tx_type_palette = character(0)
)
}
\arguments{
\item{gene_name}{Character vector of gene names to print.}

\item{gff_file}{Character string specifying the path to a GFF file
containing the gene annotation.}

\item{txdb}{A \code{TxDb} object containing the transcript database.}

\item{annotation_table}{A data frame containing the transcript annotation.}

\item{annotation_name}{Character string specifying the name of the
annotation (e.g., "Ensembl_v104").}

\item{assembly_name}{Character string specifying the name of the
genome assembly (e.g., "GRCh38").}

\item{ideogram_assembly}{Character string specifying the name of the
genome assembly for ideogram plotting (e.g., "hg38").}

\item{organism}{Character string specifying the organism name
(e.g., "Homo sapiens").}

\item{orgdb_package}{Character string specifying the name of the
organism database package (e.g., "org.Hs.eg.db").}

\item{bsgenome_package}{Character string specifying the name of the
BSgenome package (e.g., "BSgenome.Hsapiens.UCSC.hg38").}

\item{tx_type_palette}{Character vector specifying the color palette
for transcript types.}
}
\description{
The \code{ContextData} class holds information on the genomic context of
transcripts and their annotation in a \code{IsoformicExperiment} analysis.
}
\details{
The preferred way to construct an object of this class is through the
\code{\link[=create_context_data]{create_context_data()}} function.
}
\keyword{internal}
