% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_profile_data.R
\name{convert_to_isoformic_tibble}
\alias{convert_to_isoformic_tibble}
\title{Summarize Transcript-level Expression to Gene-level}
\usage{
convert_to_isoformic_tibble(txi_transcript)
}
\arguments{
\item{txi_transcript}{A \code{tibble} containing transcript-level expression abundances.}

\item{tx_to_gene}{A \code{data.frame} or \code{tibble} containing transcript-to-gene mapping information, including \code{transcript_id} and \code{gene_id} columns.}
}
\value{
A \code{tibble} containing gene-level expression abundances.
}
\description{
This function aggregates transcript-level expression data to gene-level by calculating the mean expression of transcripts belonging to the same gene.
}
\keyword{internal}
