% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContextData-class.R
\name{create_context_data}
\alias{create_context_data}
\title{Create ContextData Object}
\usage{
create_context_data(
  gff_file,
  ...,
  organism,
  orgdb_package,
  bsgenome_package,
  tx_type_palette = NULL
)
}
\arguments{
\item{gff_file}{Character string specifying the path to a GFF file
containing the gene annotation.}

\item{...}{These dots are for future extensions and must be empty.}

\item{organism}{Character string specifying the organism name
(e.g., "Homo sapiens").}

\item{orgdb_package}{Character string specifying the name of the
organism database package (e.g., "org.Hs.eg.db").}

\item{bsgenome_package}{Character string specifying the name of the
BSgenome package (e.g., "BSgenome.Hsapiens.UCSC.hg38").}

\item{tx_type_palette}{Named Character vector specifying the color palette
for transcript types.}
}
\description{
This function instantiates a \code{ContextData} object containing information
about the genomic context of transcripts.
It requires a GFF file for gene annotation and constructs a \code{TxDb}
object from it. The function also prepares an annotation table and
updates the transcript names in the \code{TxDb} object.
The \code{ContextData} object can then be used in conjunction with
\code{IsoformicExperiment} for transcriptomic analyses.
}
