% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_isoformic_mae.R
\name{create_isoformic_mae_from_se}
\alias{create_isoformic_mae_from_se}
\title{Create Isoformic Object from SummarizedExperiment}
\usage{
create_isoformic_mae_from_se(se_tx, se_gene, tx_to_gene = NULL)
}
\arguments{
\item{se_tx}{\code{SummarizedExperiment} object containing \strong{transcript-level}
differential expression analysis results.}

\item{se_gene}{\code{SummarizedExperiment} object containing \strong{gene-level}
differential expression analysis results.}
}
\description{
Create Isoformic Object from SummarizedExperiment
}
