% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_isoformic_cache.R
\name{get_isoformic_cache}
\alias{get_isoformic_cache}
\title{Retrieve System-Dependent Cache Path for Isoformic}
\usage{
get_isoformic_cache(..., ext = NULL)
}
\value{
A path character string representing the path to the user cache
directory for the \code{isoformic} package.
}
\description{
Determines the appropriate user cache directory for the \code{isoformic} package
based on the operating system.
On macOS, it avoids using paths with spaces and follows the XDG base
directory specification.
}
\details{
This function uses the \verb{[tools::R_user_dir()]} function to determine the
user cache directory.
}
