% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsoformicExperiment-class.R
\name{isoformic_experiment}
\alias{isoformic_experiment}
\title{Instantiate an IsoformicExperiment Object}
\usage{
isoformic_experiment(db_path)
}
\arguments{
\item{db_path}{Character string specifying the path to the dataset.}
}
\description{
This function creates an instance of the IsoformicExperiment class,
which is used to manage and analyze transcriptomic data.
}
\details{
It requires the path to a dataset that contains the necessary
data files, including sample metadata and annotation tables.
}
