% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsoformicExperiment-class.R
\name{IsoformicExperiment}
\alias{IsoformicExperiment}
\title{IsoformicExperiment Class}
\usage{
IsoformicExperiment(path = character(0), col_data = NULL)
}
\arguments{
\item{path}{Character string specifying the path to the dataset.}

\item{col_data}{A data frame containing sample metadata.}

\item{row_data_transcripts}{A property that retrieves transcript
annotation data.}

\item{row_data_genes}{A property that retrieves gene annotation data.}

\item{row_data_exons}{A property that retrieves exon annotation data.}

\item{row_data}{A property that aggregates transcript, gene,
and exon annotation data.}
}
\description{
The \code{IsoformicExperiment} class encapsulates the core data structure
for transcriptomic analyses in the \code{isoformic} package.
It holds the path to the dataset, sample metadata, and provides
access to transcript, gene, and exon annotations through properties.
The preferred way to construct an object of this class is through the
\code{\link[=isoformic_experiment]{isoformic_experiment()}} function.
}
\keyword{internal}
