% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_lnc_ranges.R
\name{modify_lnc_ranges}
\alias{modify_lnc_ranges}
\title{Edit the lncrna start and end table}
\usage{
  modify_lnc_ranges(lncRNA_start_end, value_to_add)
}
\arguments{
  \item{lncRNA_start_end}{Table with start and end information of the lncRNA}
  
  \item{value_to_add}{vector of a value to add on each side}
  
}
\description{
  Edit the lncrna start and end table
}