% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_plot.R
\name{plot_tx_expr}
\alias{plot_tx_expr}
\title{Plot Transcript per gene expression}
\usage{
plot_tx_expr(genes_to_plot, profile_data)
}
\arguments{
\item{genes_to_plot}{a character vector with gene names}

\item{profile_data}{tibble output from \code{prepare_profile_data}}
}
\value{
a \code{ggplot} object
}
\description{
Plot Transcript per gene expression
}
