% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_annotation_db.R
\name{prepare_annotation_db}
\alias{prepare_annotation_db}
\title{Write Parquet File from GFF}
\usage{
prepare_annotation_db(input_path, output_path = NULL, file_type = c("gff"))
}
\arguments{
\item{input_path}{Character string specifying the path to the input GFF file.}

\item{output_path}{Character string specifying the path to the output
Parquet file. If \code{NULL} or an empty string, a temporary file will be created.}

\item{file_type}{Character string specifying the type of the input file.
Currently, only "gff" is supported (default is "gff").}
}
\value{
Invisible path to the created Parquet file.
}
\description{
This function reads a GFF file and writes its contents to a Parquet file
using DuckDB.
}
