% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_row_data_from_gff.R
\name{prepare_isoformic_annotation}
\alias{prepare_isoformic_annotation}
\title{Write Feature Annotation to Parquet Files}
\usage{
prepare_isoformic_annotation(
  input_path,
  output_path = NULL,
  file_type = c("gff")
)
}
\arguments{
\item{input_path}{Character string specifying the path to the input GFF file.}

\item{output_path}{Character string specifying the path to the output
directory where Parquet files are written. If \code{NULL} or an empty string,
the cache directory will be used.}

\item{file_type}{Character string specifying the type of the input file.
Currently, only "gff" is supported (default is "gff").}
}
\value{
Invisible path to the created Parquet file.
}
\description{
This function reads an annotation file and parse feature annotation to
Parquet files each level of required
feature (i.e. gene, transcript, and exon).
}
