% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_profile_data.R
\name{prepare_profile_data}
\alias{prepare_profile_data}
\title{Prepare Data for Profile Plot}
\usage{
prepare_profile_data(
  txi_gene = NULL,
  txi_transcript,
  sample_metadata,
  tx_to_gene,
  de_result_gene,
  de_result_transcript,
  var,
  var_levels,
  gene_col = "gene_name",
  tx_col = "transcript_name",
  pvalue_cutoff = 0.05,
  lfc_cutoff = 1,
  use_fdr = TRUE
)
}
\value{
a \code{tibble} with data to plot.
}
\description{
Prepare Data for Profile Plot
}
