% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_swish.R
\name{run_swish_pairwise}
\alias{run_swish_pairwise}
\title{Run Swish Differential Expression}
\usage{
run_swish_pairwise(se, contrast_var = "condition", ...)
}
\arguments{
\item{se}{\code{SummarizedExperiment} object.}

\item{contrast_var}{column name from \code{colData(se)}.}

\item{...}{Additional arguments passed to \code{fishpond::swish()}.}
}
\value{
\code{SummarizedExperiment} object.
}
\description{
Run swish method for a \code{SummarizedExperiment} with inferential replicates.
}
