% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salmon_utils.R
\name{salmon_index}
\alias{salmon_index}
\title{Build Salmon Index}
\usage{
salmon_index(
  fasta_path,
  index_path = "salmon_index",
  kmer_len = 31,
  num_threads = 2,
  env_name = "salmon-env",
  is_gencode = FALSE,
  decoy_fasta = NULL,
  clip_poly_a = TRUE
)
}
\arguments{
\item{fasta_path}{Path to the reference transcriptome FASTA file.}

\item{index_path}{Directory path to save the Salmon index (default is
"salmon_index").}

\item{kmer_len}{K-mer length for the index (default is 31).}

\item{num_threads}{Number of threads to use (default is 2).}

\item{env_name}{Name of the conda environment with Salmon installed
(default is "salmon-env").}

\item{is_gencode}{Logical indicating if the FASTA is from GENCODE
(default is \code{FALSE}).}

\item{decoy_fasta}{Optional path to a FASTA file containing decoy sequences
(default is \code{NULL}).}

\item{clip_poly_a}{Logical indicating whether to clip poly-A tails
(default is \code{TRUE}).}
}
\value{
processx style output list.
}
\description{
Create a Salmon index from a reference transcriptome FASTA file.
}
