% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_utils.R
\name{se_experiment_level}
\alias{se_experiment_level}
\title{Return \code{SummarizedExperiment} Experiment Level}
\usage{
se_experiment_level(se)
}
\arguments{
\item{se}{object of class \code{SummarizedExperiment}.}

\item{type}{String containing one of \code{c("txp", "gene")}.}
}
\description{
Check level of quantification for
}
