% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_isoformic_mae.R
\name{validate_isoformic_mae}
\alias{validate_isoformic_mae}
\title{Validate Isoformic Object}
\usage{
validate_isoformic_mae(mae)
}
\arguments{
\item{mae}{\code{MultiAssayExperiment} object to be validated.}
}
\value{
\code{MultiAssayExperiment} object with valid \code{{isoformic}} slots.
}
\description{
Validate if \code{MultiAssayExperiment} object is compatible with\code{{isoformic}}.
Also, validates if \code{{isoformic}} preparation steps have been executed
for the object.
}
