% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_reference.R
\name{download_reference}
\alias{download_reference}
\title{Download Reference Files From GENCODE}
\usage{
download_reference(
  version = "43",
  reference = "gencode",
  file_type = "gff",
  output_path = "data-raw",
  timeout_limit = 3600,
  method = "auto"
)
}
\arguments{
\item{file_type}{One of gff, fasta, or gtf.}
}
\description{
Download Reference Files From GENCODE
}
