% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{is_deg_sig}
\alias{is_deg_sig}
\title{Add Differential Gene Expression Results to Transcript Table}
\usage{
is_deg_sig(DegsigVector, DET_table)
}
\description{
Add Differential Gene Expression Results to Transcript Table
}
