% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoformic-package.R
\docType{package}
\name{isoformic-package}
\alias{isoformic}
\alias{isoformic-package}
\title{isoformic: Isoform-Level Biological Interpretation of Transcriptomic Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Isoform-level biological interpretation of transcriptomic data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/luciorq/isoformic}
  \item \url{https://luciorq.github.io/isoformic/}
  \item Report bugs at \url{https://github.com/luciorq/isoformic/issues}
}

}
\author{
\strong{Maintainer}: Lucio Rezende Queiroz \email{luciorqueiroz@gmail.com, lrq4001@med.cornell.edu} (\href{https://orcid.org/0000-0002-6090-1834}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Lucio Rezende Queiroz \email{luciorqueiroz@gmail.com, lrq4001@med.cornell.edu} (\href{https://orcid.org/0000-0002-6090-1834}{ORCID}) [copyright holder]
  \item Izabela Mamede Conceicao \email{iza.mamede@gmail.com} (\href{https://orcid.org/0000-0002-0707-5588}{ORCID}) [contributor]
  \item Carlos Mata-Machado \email{carlosaugustodmm@gmail.com} (\href{https://orcid.org/0009-0005-6615-6563}{ORCID}) [contributor]
  \item Luigi Marchionni \email{lum4003@med.cornell.edu} (\href{https://orcid.org/0000-0002-7336-8071}{ORCID}) [contributor]
  \item Gloria Franco \email{gfranco@ufmg.br} (\href{https://orcid.org/0000-0001-5245-2365}{ORCID}) [contributor]
}

}
\keyword{internal}
