% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{join_DEG_DET}
\alias{join_DEG_DET}
\title{Merge Gene and Transcript Level Differential Expression Tables}
\usage{
join_DEG_DET(DEG_tab, DET_final_tab, logfc_cut, pval_cut)
}
\description{
Merge Gene and Transcript Level Differential Expression Tables
}
