% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_tx_context}
\alias{plot_tx_context}
\title{Plot Genomic Context From GTF Annotation File}
\usage{
plot_tx_context(exon_table)
}
\arguments{
\item{exon_table}{Exon table generated from \code{prepare_exon_annotation}}
}
\description{
Plot Genomic Context From GTF Annotation File
}
