% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_exon_annotation.R
\name{prepare_exon_annotation}
\alias{prepare_exon_annotation}
\title{Prepare Exon based Position Annotation Table}
\usage{
prepare_exon_annotation(gene_name, file_path, file_type = c("gff", "gtf"))
}
\arguments{
\item{gene_name}{String or vector of gene names to extract.}

\item{file_path}{Path to annotation file.}

\item{file_type}{A character string specifying the type of file to download.
Valid options are \code{"gff"}, \code{"gtf"}, \code{"fasta"} or \code{"genome_fasta"}.
Defaults to \code{"gff"}.
\strong{Note:} \code{"fasta"} refers to the transcriptome FASTA file.
\code{"genome_fasta"} refers to the whole genome sequence FASTA file.}
}
\description{
Prepare Exon based Position Annotation Table
}
