% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_exon_annotation.R
\name{prepare_exon_annotation}
\alias{prepare_exon_annotation}
\title{Prepare Exon based Position Annotation Table}
\usage{
prepare_exon_annotation(gene_name, file_path, file_type = c("gff"))
}
\arguments{
\item{file_path}{Path to annotation file.}

\item{file_type}{Character indicating the type of file to download.
One of \code{"gtf"}, \code{"gff"}, or \code{"fasta"}. Defaults to \code{"gtf"}.
\strong{NOTE:} \code{"fasta"} refers to the transcriptome FASTA.}

\item{gene_names}{String or vector of gene names to extract.}
}
\description{
Prepare Exon based Position Annotation Table
}
